requirements.txt
Django>=4.2,<5
stripe>=10.0,<12
gunicorn>=22.0,<24
python-dotenv>=1.0,<2
whitenoise>=6.6,<7
